Week13/board/src/test/java/com/nhnacademy/jdbc/board/controller/HomeControllerTest.java
package com.nhnacademy.jdbc.board.controller;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;

class HomeControllerTest {

    MockMvc mockMvc;

    @BeforeEach
    void setUp() {
        mockMvc = MockMvcBuilders.standaloneSetup(HomeController.class)
                .build();
    }

    @Test
    void index() throws Exception {
        mockMvc.perform(get("/"))
                .andExpect(status().isOk())
                .andExpect(view().name("index/index"))
                .andDo(print());
    }

    @Test
    void noAccess() throws Exception {
        mockMvc.perform(get("/nonAccess"))
                .andExpect(status().isOk())
                .andExpect(view().name("noAccess"))
                .andDo(print());
    }
}
